rename ibis.py to _ibis.py
