+++ vdsql/__init__.py
@@ -1,3 +1,3 @@
 from .__about__ import *
-from .ibis import *
+from ._ibis import *
 from .bigquery import *
